♻️ refactor: 레시피 페이지 수정
